feat: waits for last publishing before shutdown
1) breaks out publisher, so it can be extended by implementations
2) adds `config.publisher` for a custom Publisher in registry
3) waits for optional callback in `registry.stop` and `publisher`
4) replace handcrafted http client with will maintained lean module

NOTE: this commit adds two lean depedencies to spectator-js
